refactor: 💡 organize the cookie related code
✅ Closes: #1343
