+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "src",
-  "version": "0.8.0+64.2",
+  "version": "0.8.0+64.3",
   "private": true,
   "scripts": {
     "dev": "next dev",
